+++ src/authenticationProvider/authenticationProvider.js
@@ -7,7 +7,9 @@ import {
 } from "@inrupt/solid-client";
 import { getDefaultSession, fetch } from "@inrupt/solid-client-authn-browser";
 import { FOAF } from "@inrupt/vocab-common-rdf";
-import SparqlDataProvider from "./../dataProvider/SparqlDataProvider";
+
+import comunicaEngineWrapper from "../comunicaEngineWrapper/comunicaEngineWrapper";
+
 
 const queryEngine = new QueryEngine();
 
@@ -40,7 +42,7 @@ export default {
     }
   },
   logout: async function logout() {
-    await SparqlDataProvider.queryEngine.invalidateHttpCache();
+    await comunicaEngineWrapper.reset();
     await queryEngine.invalidateHttpCache();
     const session = getDefaultSession();
     await session.logout();